+++ Changelog.md
@@ -12,3 +12,4 @@ Added back Sodium and Iris this patch since the zipping of the MI datagen files 
 - [Added the textures and models that were missing for clear glass](https://github.com/TheStaticVoid/StaTech-Industry/issues/449)
 - [Fix Structory and Structory Towers being the forge version](https://github.com/TheStaticVoid/StaTech-Industry/commit/2895a52951491c65db9ec5320468694c7e9cfeb7) - packwiz does a little trolling
 - [Added Dragons Blood to the Blood Generator](https://github.com/TheStaticVoid/StaTech-Industry/issues/452)
+- [Added recipe to make Netherrack from Cobblestone and Blood in a Mixer](https://github.com/TheStaticVoid/StaTech-Industry/issues/453)